requirements.txt
google-cloud-documentai>=3.1.0  # Latest stable release (v3.1.0) :cite[1]
openai>=1.60.1                 # Latest release (v1.60.1) :cite[2]:cite[4]
python-dotenv>=1.0.1           # Latest stable version (v1.0.1) :cite[3]:cite[6]